restructure project with cmake
