+++ src/models/eventData.hpp
@@ -1,0 +1,44 @@
+
+#pragma once
+
+#include "eventDataRow.hpp"
+#include <functional>
+#include <vector>
+
+class EventData {
+public:
+    std::chrono::system_clock::time_point deliveryStart;
+    std::chrono::system_clock::time_point deliveryEnd;
+    std::string product;
+    std::vector<EventDataRow> rows;
+
+    // Constructor with all parameters including data
+    EventData(
+        std::chrono::system_clock::time_point start,
+        std::chrono::system_clock::time_point end,
+        std::string prod,
+        std::vector<EventDataRow> rows);
+
+    // Constructor without data parameter
+    EventData(
+        std::chrono::system_clock::time_point start,
+        std::chrono::system_clock::time_point end,
+        std::string prod);
+
+    void addRow(EventDataRow row); 
+
+    EventData filter(std::function<bool(const EventDataRow&)> predicate);
+
+    void forEach(std::function<void(const EventDataRow&)> func);
+
+    std::string toString() const;
+
+private:
+    static std::string timePointToString(const std::chrono::system_clock::time_point& tp);
+
+    static std::string sideToString(Side side);
+
+    static std::string executionRestrictionToString(ExecutionRestriction er);
+
+    static std::string actionCodeToString(ActionCode ac);
+};
