+++ src/orderbook/orderbook.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include "models/eventData.hpp"
+
+EventData getLOB(EventData eventData, std::chrono::system_clock::time_point time_stamp);
