+++ src/utils/csv.hpp
@@ -1,0 +1,4 @@
+
+#include "models/eventData.hpp"
+
+EventData loadEventDataFromCSV(const std::string& filename, bool ignoreFirstLine);
